moving config to the root folder
